+++ rpcs3/Emu/RSX/NV47/HW/nv4097.cpp
@@ -20,8 +20,29 @@ namespace rsx
 			RSX(ctx)->m_graphics_state |= rsx::pipeline_state::fragment_program_ucode_dirty;
 		}
 
-		void set_transform_constant::impl(context* ctx, u32 reg, u32 /*arg*/)
+		void set_transform_constant::decode_one(context* ctx, u32 reg, u32 arg)
 		{
+			const u32 index = reg - NV4097_SET_TRANSFORM_CONSTANT;
+			const u32 constant_id = index / 4;
+			const u8 subreg = index % 4;
+			const u32 load = REGS(ctx)->transform_constant_load();
+
+			REGS(ctx)->transform_constants[load + constant_id][subreg] = arg;
+		}
+
+		void set_transform_constant::impl(context* ctx, u32 reg, u32 arg)
+		{
+			if (RSX(ctx)->in_begin_end && !REGS(ctx)->current_draw_clause.empty())
+			{
+				// Updating constants mid-draw is messy. Push attr barrier.
+				REGS(ctx)->current_draw_clause.insert_command_barrier(
+					rsx::transform_constant_update_barrier,
+					arg,
+					reg - NV4097_SET_TRANSFORM_CONSTANT
+				);
+				return;
+			}
+
 			const u32 index = reg - NV4097_SET_TRANSFORM_CONSTANT;
 			const u32 constant_id = index / 4;
 			const u8 subreg = index % 4;
@@ -437,6 +458,11 @@ namespace rsx
 
 			// Rollback
 			REGS(ctx)->decode(reg, REGS(ctx)->latch);
+		}
+
+		void set_transform_constant_load(context* ctx, u32 reg, u32 arg)
+		{
+			util::push_draw_parameter_change(ctx, rsx::transform_constant_load_modifier_barrier, reg, arg);
 		}
 
 		///// Reports
